% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{annotate_splits}
\alias{annotate_splits}
\title{split the text document to run the model parallelly}
\usage{
annotate_splits(x, file, individuals = TRUE)
}
\arguments{
\item{x}{text variable}

\item{file}{provide the directory path to udpipe model}
}
\description{
split the text document to run the model parallelly
}
\examples{
future_lapply(X = corpus_splitted, FUN = annotate_splits, file = ud_model$file, future.seed = TRUE) \%
}
