% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_poisIRT}
\alias{bootstrap_poisIRT}
\title{Parametric Bootstrap of EM Standard Errors}
\usage{
bootstrap_poisIRT(
  emIRT.out,
  .data,
  .starts,
  .priors,
  .control = {     list(threads = 1, verbose = TRUE, thresh = 1e-06, maxit = 5000) },
  set.seed = 1234,
  Ntrials = 5
)
}
\arguments{
\item{emIRT.out}{an emIRT() object, which is output from a call to poisIRT}

\item{.data}{the data used to produce the emIRT object.}

\item{.starts}{a list containing several matrices of starting values for the parameters, which include alpha A (J x 1) matrix , psi A (K x 1) matrix, beta A (J x 1) matrix and x An (NI x 1) matrix.}

\item{.priors}{list, containing several matrices of starting values for the parameters. .priors can be generated by create_prior() by default via mu = 0 and variance = 100.}
}
\description{
bootstrap_poisIRT take an poisIRT() object (from emIRT package) and implements a parametric bootstrap of the standard errors for the ideal points.
It assumes you have already run the Poisson IRT estimation via EM , and takes the output from that estimate,
along with the original dataset, word frequency matrix. This function will conduct a bootstrap by running the estimates from sub-sampled observations based on the a poisson distribution.
}
