% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{conll}
\alias{conll}
\title{Pre-processed textual files parsed on CoNLL-U format}
\format{
A data frame with 831,639 rows and 14 variables
}
\usage{
data(conll)
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds. The variables are as follows:
}
\keyword{CoNLL}
