% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_prior}
\alias{create_prior}
\title{create prior variance and meean  for poisIRT}
\usage{
create_prior(mu = 0, sigma2 = 100, ...)
}
\arguments{
\item{mu}{prior mean for x_i, β_j, ψ_k and ψ_k.}

\item{sigma2}{prior variance for x_i, β_j, ψ_k and ψ_k.}
}
\description{
create prior variance and meean  for poisIRT
}
\examples{
create_prior(mu = 0, sigma2 = 100)
}
