% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_start}
\alias{create_start}
\title{create start point for poisIRT()}
\usage{
create_start(wfm_matrix)
}
\arguments{
\item{wfm_matrix}{wfm object}
}
\description{
create start point for poisIRT()
}
\examples{
create_start(wfm)
}
