% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_dictionary}
\alias{get_dictionary}
\title{Generate dictionary object via a matrix or data frame}
\usage{
get_dictionary(df)
}
\arguments{
\item{folder}{a data frame or matrix}
}
\description{
Generate dictionary object via a matrix or data frame
}
\examples{
get_dictionary(mat)
}
