% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_keywords}
\alias{get_keywords}
\title{retrieve the most/less frequent keywords from textrank_keywords class}
\usage{
get_keywords(df, n = 10, head = TRUE, ...)
}
\arguments{
\item{df}{the estimate object should be textrank_keywords class from textrank}
}
\description{
retrieve the most/less frequent keywords from textrank_keywords class
}
