% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_wordfeatures}
\alias{get_wordfeatures}
\title{retrieve word features from poisIRT class}
\usage{
get_wordfeatures(df, ...)
}
\arguments{
\item{df}{the estimate object from poisIR}
}
\description{
retrieve word features from poisIRT class
}
