% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_tag.R
\name{pos_tag}
\alias{pos_tag}
\title{First Step: Parsing Simplified Chinese Text with udpipe in R and 2.5 Chinese-GSDSIM Model}
\usage{
pos_tag(dataframe, self_language_model = FALSE, model_name, ...)
}
\arguments{
\item{dataframe}{A \code{dataframe} at least should contain \code{text variable} and \code{doc_id}.}

\item{self_language_model}{False}

\item{model_name}{You can provide your trained model but you need to locate}
}
\description{
This function call works on udpipe packages and parallel computing packages in R.
}
\references{

}
