% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{package}
\name{redgaurds}
\alias{redgaurds}
\alias{.onAttach}
\title{redgaurds: A package for replicating  estimates and findings in the article of Factionalism and the
Red Guards under Mao's China: Ideal Point Estimation Using Text Data.}
\usage{
.onAttach(...)
}
