% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start-replication.R
\name{run_replication}
\alias{run_replication}
\title{Run replication}
\usage{
run_replication(folder = "replication-code")
}
\arguments{
\item{folder}{the folder name for restoring the relication r scrips}
}
\description{
Run replication
}
\examples{
run_replication()
}
