% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{annotate_splits}
\alias{annotate_splits}
\title{split the text document to run the model in parallel}
\usage{
annotate_splits(
  x,
  file = "chinese-gsdsimp-ud-2.5-191206.udpipe",
  individuals = TRUE
)
}
\arguments{
\item{x}{text variable}

\item{file}{provide the directory path to udpipe model}
}
\description{
split the text document to run the model in parallel
}
\examples{
annotate_splits()
}
