% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_start}
\alias{create_start}
\title{create start point for poisIRT()}
\usage{
create_start(x, set.seed = 1234, verbose = FALSE)
}
\arguments{
\item{x}{wfm object}

\item{set.seed}{default is 1234}

\item{verbose}{default is FALSE}
}
\description{
create start point for poisIRT()
}
\examples{
create_start(data)
}
