% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dtm_wfm}
\alias{dtm_wfm}
\title{Turning DTM to WFM}
\usage{
dtm_wfm(dtm_object)
}
\arguments{
\item{dtm_object}{input requires dtm object from Quanteda}
}
\description{
Turning DTM to WFM
}
\examples{
dtm_wfm(dtm)
}
