% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_estimates}
\alias{get_estimates}
\title{retrieve estimates from poisIRT class}
\usage{
get_estimates(df)
}
\arguments{
\item{folder}{A table of weather variables with PWS created by hist_data function}
}
\description{
retrieve estimates from poisIRT class
}
\examples{
get_estimates(poisIRT_object)
}
