% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_estimates}
\alias{get_estimates}
\title{retrieve estimates from poisIRT class}
\usage{
get_estimates(emIRT.out)
}
\arguments{
\item{emIRT.out}{the estimate object from poisIR}
}
\description{
retrieve estimates from poisIRT class
}
