% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pos_tagging}
\alias{pos_tagging}
\title{part-of-speech at individual level or incident level}
\usage{
pos_tagging(df, individuals = TRUE)
}
\arguments{
\item{df}{text document object}
}
\description{
part-of-speech at individual level or incident level
}
\details{
This function requires pre-trianed model (chinese-gsdsimp-ud-2.5-191206.udpipe) which can be downloaded at Pre-trained models: chinese-gsdsimp-ud-2.5-191206.udpipe.
}
\note{
This is limited to used for specific data frame from replicaiton dataset
}
\examples{
pos_tagging(incindent)
}
